Move discovery streaming from 'mngr list --stream' to 'mngr observe --discovery-only'
The streaming functionality doesn't belong in 'list' because none of the
list filtering/sorting options make sense with it. It fits naturally as
a mode of 'observe' since observe already runs it internally.

- Extract discovery stream logic from cli/list.py to api/discovery_events.py
  as run_discovery_stream() with an optional on_line callback
- Add --discovery-only flag to 'mngr observe' CLI
- Remove --stream flag from 'mngr list' CLI
- Update AgentObserver to spawn 'mngr observe --discovery-only' subprocess
- Update MngrStreamManager in minds app to use new command
- Move stream tests to discovery_events_test.py, update list tests
- Regenerate CLI docs, update tutorial examples
- Fix pre-existing ruff issues in observe.py (unused pdb import, B007)
