+++ libs/mngr/docs/commands/primary/list.md
@@ -52,12 +52,11 @@ mngr list [OPTIONS]
 | `--sort` | text | Sort by CEL expression(s) with optional direction, e.g. 'name asc, create_time desc'; enables sorted (non-streaming) output [default: create_time] | `create_time` |
 | `--limit` | integer | Limit number of results (applied after fetching from all providers) | None |
 
-## Watch / Stream Mode
+## Watch Mode
 
 | Name | Type | Description | Default |
 | ---- | ---- | ----------- | ------- |
 | `-w`, `--watch` | integer | Continuously watch and update status at specified interval (seconds) | None |
-| `--stream` | boolean | Stream discovery events as JSONL. Outputs a full snapshot, then tails the event file for updates. Periodically re-polls to catch any missed changes. | `False` |
 
 ## Error Handling
