+++ libs/mngr/docs/commands/secondary/observe.md
@@ -6,7 +6,7 @@
 **Synopsis:**
 
 ```text
-mngr observe [--events-dir DIR]
+mngr observe [--events-dir DIR] [--discovery-only]
 ```
 
 Observe agent state changes across all hosts [experimental].
@@ -19,7 +19,7 @@ events to local JSONL files:
 
 The observer:
 1. Loads base state from event history (if available) to detect state changes since last run
-2. Uses 'mngr list --stream' to track which hosts are online
+2. Runs host discovery to track which hosts are online
 3. Streams activity events from each online host
 4. When activity is detected, fetches and emits agent state for the affected host
 5. Periodically (every 5 minutes) emits a full state snapshot of all agents
@@ -27,6 +27,10 @@ The observer:
 Only one instance per output directory can run at a time (enforced via file lock).
 Use --events-dir to write events to a different directory, allowing multiple
 observers to run simultaneously for different output locations.
+
+With --discovery-only, only the host/agent discovery stream is emitted as JSONL
+to stdout. This is useful for programmatically tracking which agents and hosts
+exist without the full observe overhead.
 
 Press Ctrl+C to stop.
 
@@ -64,6 +68,8 @@ mngr observe [OPTIONS]
 | Name | Type | Description | Default |
 | ---- | ---- | ----------- | ------- |
 | `--events-dir` | path | Base directory for event output files and lock. Defaults to MNGR_HOST_DIR (~/.mngr). | None |
+| `--discovery-only` | boolean | Stream only discovery events as JSONL (hosts and agents discovered/destroyed). Outputs a full snapshot, then tails the event file for updates. Periodically re-polls to catch any missed changes. Does not start activity streams or emit agent state events. | `False` |
+| `--on-error` | choice (`abort` &#x7C; `continue`) | What to do when errors occur: abort (stop immediately) or continue (keep going) | `abort` |
 
 ## See Also
 
@@ -88,4 +94,10 @@ $ mngr observe --events-dir /path/to/events
 
 ```bash
 $ mngr observe --quiet
+```
+
+**Stream only discovery events**
+
+```bash
+$ mngr observe --discovery-only
 ```
